Refactor Model
